% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric.R
\name{arc_length}
\alias{arc_length}
\title{The length of an arc}
\usage{
arc_length(radius, theta)
}
\arguments{
\item{radius}{The radius of a circle}

\item{theta}{the angle which the arc subtends at the centre of the circle, in radians}
}
\description{
The length of an arc
}
