% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric.R
\name{cosecant}
\alias{cosecant}
\title{cosecant}
\usage{
cosecant(r, y)
}
\arguments{
\item{r}{radius or hypotenuse of an acute angle in terms of a right triangle}

\item{y}{y coordinate in a circle or opposite of an acute angle in terms of a right triangle}
}
\description{
cosecant
}
