% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric.R
\name{cosine}
\alias{cosine}
\title{cosine}
\usage{
cosine(x, r)
}
\arguments{
\item{x}{x coordinate in a circle or adjacent of an acute angle in terms of a right triangle}

\item{r}{radius or hypotenuse of an acute angle in terms of a right triangle}
}
\description{
cosine
}
