% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric.R
\name{cotangent}
\alias{cotangent}
\title{cotangent}
\usage{
cotangent(x, y)
}
\arguments{
\item{x}{x coordinate in a circle or adjacent of an acute angle in terms of a right triangle}

\item{y}{y coordinate in a circle or opposite of an acute angle in terms of a right triangle}
}
\description{
cotangent
}
