% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summation.R
\name{deviation_scores}
\alias{deviation_scores}
\title{Deviation Scores}
\usage{
deviation_scores(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vectors whose deviation scores are wanted}

\item{na.rm}{logical; if true, any \code{\link{NA}} and \code{NaN}'s
    are removed from \code{x} before the quantiles are computed.}
}
\description{
Deviation Scores
}
\details{
\emph{Deviation Scores} are just the difference between each observation and the sample mean
}
