% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summation.R
\name{idealf}
\alias{idealf}
\title{Compute quartiles based on ideal fourths}
\usage{
idealf(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector whose quantiles are wanted.}

\item{na.rm}{logical; if true, any \code{\link{NA}} and \code{NaN}'s
    are removed from \code{x} before the quantiles are computed.}
}
\value{
The lower quantile and upper quantile
}
\description{
Compute quartiles based on ideal fourths
}
\details{
The sample median divides the data into two parts: the lower half and the upper half after putting the observations in ascending order.

 \emph{Quartiles} are measures of location aimed at dividing data into four parts. This is done with two additional measures of location called the lower and upper quartiles. The median is sometimes called the middle quartile.
 
  Roughly, the \emph{lower quartile} is the median of the smaller half of the data and the \emph{upper quartile} is the median of the upper half.
}
