% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summation.R
\name{idealfIQR}
\alias{idealfIQR}
\title{Compute the interquartile range using the ideal fourths}
\usage{
idealfIQR(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector whose quantiles are wanted.}

\item{na.rm}{logical; if true, any \code{\link{NA}} and \code{NaN}'s
    are removed from \code{x} before the quantiles are computed.}
}
\description{
Compute the interquartile range using the ideal fourths
}
\seealso{
\code{\link{idealf}}
}
