% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summation.R
\name{interquartile_range}
\alias{interquartile_range}
\title{Interquartile Range}
\usage{
interquartile_range(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vectors whose interquartile range is wanted}

\item{na.rm}{logical; if true, any \code{\link{NA}} and \code{NaN}'s
    are removed from \code{x} before the quantiles are computed.}
}
\description{
Interquartile Range
}
\details{
For some purposes, it is important to measure the variability of the centrally located values.

One approach is the interquartile range, which is just \code{qu − ql} , the difference between the upper and lower quartiles.

Notice that the interquartile range is insensitive to the more extreme values.  The upper and lower quartiles are resistant to outliers, which means that the most extreme values do not affect the values of \code{qu} and \code{ql}. Consequently, the interquartile range is resistant to outliers as well.
}
