% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summation.R
\name{mad}
\alias{mad}
\title{Median Absolute Deviation}
\usage{
mad(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vectors whose median absolute deviation is wanted}

\item{na.rm}{logical; if true, any \code{\link{NA}} and \code{NaN}'s
    are removed from \code{x} before the quantiles are computed.}
}
\description{
Median Absolute Deviation
}
\details{
MAD is computed by subtracting the median from each observation, taking the absolute value of each difference, and then computing the median.

Typically, \href{https://en.wikipedia.org/wiki/Median_absolute_deviation}{MAD} is divided by 0.6745. 
MADN = MAD / 0.6745
}
\examples{
x <- c(12,45,23,79,19,92,30,58,132)
mad(x)
}
