% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summation.R
\name{madn}
\alias{madn}
\title{Median Absolute Deviation N}
\usage{
madn(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vectors whose median absolute deviation is wanted}

\item{na.rm}{logical; if true, any \code{\link{NA}} and \code{NaN}'s
    are removed from \code{x} before the quantiles are computed.}
}
\description{
Median Absolute Deviation N
}
\seealso{
\code{\link{mad}}
}
