% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summation.R
\name{outms}
\alias{outms}
\title{Classic Outlier Detection Method}
\usage{
outms(x, crit = 2, na.rm = TRUE, plotit = TRUE)
}
\arguments{
\item{x}{An \R object.  Currently there are methods for
    numeric/logical vectors and \link[=Dates]{date},
    \link{date-time} and \link{time interval} objects.  Complex vectors
    are allowed for \code{trim = 0}, only.}

\item{crit}{how many standard deviations from the mean?}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}

\item{plotit}{Logical; Plot it or not?}
}
\description{
Classic Outlier Detection Method
}
\details{
In words, declare X an outlier if it is more than two standard deviations from the mean.

A commonly used variation of this rule is to declare a value an outlier if it is more than three standard deviations from the mean.
}
\examples{
x <- c(-158, 2, 8, 9, 10, 16, 18, -148, 22, 27, 29, 35, 42)
outms(x, plotit = TRUE)
}
