% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summation.R
\name{outpro}
\alias{outpro}
\title{Detect outlier using the MAD–Median Rule}
\usage{
outpro(x, na.rm = TRUE, plotit = TRUE)
}
\arguments{
\item{x}{An \R object.  Currently there are methods for
    numeric/logical vectors and \link[=Dates]{date},
    \link{date-time} and \link{time interval} objects.  Complex vectors
    are allowed for \code{trim = 0}, only.}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}

\item{plotit}{Logical; Plot it or not?}
}
\description{
Detect outlier using the MAD–Median Rule
}
