% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summation.R
\name{sample_standard_deviation}
\alias{sample_standard_deviation}
\title{Sample Standard Deviation}
\usage{
sample_standard_deviation(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or an \R object which is coercible to one
    by \code{as.double(x)}.}

\item{na.rm}{logical.  Should missing values be removed?}
}
\description{
Sample Standard Deviation
}
\details{
The sample standard deviation is the (positive) square root of the variance
}
