% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summation.R
\name{sample_variance}
\alias{sample_variance}
\title{Sample Variance}
\usage{
sample_variance(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
Sample Variance
}
