% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric.R
\name{secant}
\alias{secant}
\title{secant}
\usage{
secant(r, x)
}
\arguments{
\item{r}{radius or hypotenuse of an acute angle in terms of a right triangle}

\item{x}{x coordinate in a circle or adjacent of an acute angle in terms of a right triangle}
}
\description{
secant
}
