% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric.R
\name{sine}
\alias{sine}
\title{sine}
\usage{
sine(y, r)
}
\arguments{
\item{y}{y coordinate in a circle or opposite of an acute angle in terms of a right triangle}

\item{r}{radius or hypotenuse of an acute angle in terms of a right triangle}
}
\description{
sine
}
