% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric.R
\name{tangent}
\alias{tangent}
\title{tangent}
\usage{
tangent(y, x)
}
\arguments{
\item{y}{y coordinate in a circle or opposite of an acute angle in terms of a right triangle}

\item{x}{x coordinate in a circle or adjacent of an acute angle in terms of a right triangle}
}
\description{
tangent
}
