% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric.R
\name{to_degree}
\alias{to_degree}
\title{Convert radian to degree}
\usage{
to_degree(theta)
}
\arguments{
\item{theta}{angle in radians}
}
\description{
Convert radian to degree
}
\details{
1 radian = 180/pi (about 57.3) degrees
}
