% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometric.R
\name{to_radian}
\alias{to_radian}
\title{Convert degree to radian}
\usage{
to_radian(theta)
}
\arguments{
\item{theta}{angle in degrees}
}
\description{
Convert degree to radian
}
\details{
1 degree = pi / 180 (about 0.017) radians
}
