% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summation.R
\name{trimmed_mean}
\alias{trimmed_mean}
\title{Trimmed Mean}
\usage{
trimmed_mean(x, trim = 0.2, na.rm = FALSE)
}
\arguments{
\item{x}{An \R object.  Currently there are methods for
    numeric/logical vectors and \link[=Dates]{date},
    \link{date-time} and \link{time interval} objects.  Complex vectors
    are allowed for \code{trim = 0}, only.}

\item{trim}{the fraction (0 to 0.5) of observations to be
    trimmed from each end of \code{x} before the mean is computed.
    Values of trim outside that range are taken as the nearest endpoint.
  }

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}
}
\description{
The sample median is an example of what is called atrimmed mean; it trims all but one or two values. The sample mean represents the other extreme: zero trimming.

 No specific amount of trimming is always best, but for various reasons, 20\% trimming is often a good choice. This means that the smallest 20\% and the largest 20\% are trimmed and the average of the remaining data is computed.

 In symbols, first compute 0.2n, round down to the nearest integer, call this result g, in which case the 20\% trimmed mean is given by

 \deqn{\overline{X} =  \frac{1}{n - 2g} (X_{(g+1)} + \cdots + X_{(n-g)})}
}
