% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summation.R
\name{winsd}
\alias{winsd}
\title{Winsorized sample standard deviation}
\usage{
winsd(x, trim = 0.2, na.rm = FALSE)
}
\arguments{
\item{x}{An \R object.  Currently there are methods for
    numeric/logical vectors and \link[=Dates]{date},
    \link{date-time} and \link{time interval} objects.  Complex vectors
    are allowed for \code{trim = 0}, only.}

\item{trim}{the fraction (0 to 0.5) of observations to be
    trimmed from each end of \code{x} before the mean is computed.
    Values of trim outside that range are taken as the nearest endpoint.
  }

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}
}
\description{
Winsorized sample standard deviation
}
\examples{
x <- c(1, 2, 8, 9, 10, 16, 18, 22, 27, 29, 35, 42)
winsd(x)
}
