% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summation.R
\name{winvar}
\alias{winvar}
\title{Winsorized variance}
\usage{
winvar(x, trim = 0.2, na.rm = FALSE)
}
\arguments{
\item{x}{An \R object.  Currently there are methods for
    numeric/logical vectors and \link[=Dates]{date},
    \link{date-time} and \link{time interval} objects.  Complex vectors
    are allowed for \code{trim = 0}, only.}

\item{trim}{the fraction (0 to 0.5) of observations to be
    trimmed from each end of \code{x} before the mean is computed.
    Values of trim outside that range are taken as the nearest endpoint.
  }

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}
}
\description{
Winsorized variance
}
\details{
When using a trimmed mean, certain types of analyses, require a Winsorized variance. 

The process of Winsorizing data by 20\% is related to 20\% trimming. Recall that when computing a 20\% trimmed mean, the \code{g} smallest and the \code{g} largest observations are removed, where \code{g = 0.2n} rounded down to the nearest integer, in which case the 20\% trimmed mean is the average of the remaining values.

Winsorizing the data by 20\% means that the \code{g} smallest values are not trimmed, but rather, they are set equal to the smallest value not trimmed. Similarly, the \code{g} largest values are set equal to the largest value not trimmed.
}
\examples{
x <- c(1, 2, 8, 9, 10, 16, 18, 22, 27, 29, 35, 42)
winvar(x)
}
